+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v1.48.2
-- **[fix]** - color `TypographySkeleton` disabled text
+## v1.48.3
+- **[chore]** - upgrade storybook to 7.5.3
+- **[fix]** - `Button` loading with startIcon & endIcon